invert prompt gate (enabled -> disabled) and move adminGate back to src/tools/web/
