+++ src/prompts/staleContent/inform.test.ts
@@ -7,12 +7,12 @@ describe('stale-content-cleanup-inform prompt', () => {
     expect(prompt.name).toBe('stale-content-cleanup-inform');
   });
 
-  it('is enabled only when adminToolsEnabled is true', () => {
+  it('is disabled when adminToolsEnabled is false', () => {
     const prompt = getStaleContentCleanupInformPrompt(new WebMcpServer());
 
-    expect(prompt.enabled({ adminToolsEnabled: true } as any)).toBe(true);
+    expect(prompt.disabled({ adminToolsEnabled: true } as any)).toBe(false);
 
-    expect(prompt.enabled({ adminToolsEnabled: false } as any)).toBe(false);
+    expect(prompt.disabled({ adminToolsEnabled: false } as any)).toBe(true);
   });
 
   it('instructs the model to call get-stale-content-report once and forbid recomputation', async () => {
