+++ src/tools/web/extractRefreshTasks/listExtractRefreshTasks.test.ts
@@ -1,5 +1,4 @@
 import { CallToolResult } from '@modelcontextprotocol/sdk/types.js';
-import { Ok } from 'ts-results-es';
 
 import { WebMcpServer } from '../../../server.web.js';
 import invariant from '../../../utils/invariant.js';
@@ -32,7 +31,10 @@ vi.mock('../../../restApiInstance.js', () => ({
 }));
 
 vi.mock('../adminGate.js', () => ({
-  assertAdmin: mocks.mockAssertAdmin,
+  adminGate: {
+    assertAdmin: mocks.mockAssertAdmin,
+    clearCache: vi.fn(),
+  },
 }));
 
 vi.mock('../../../config.js', () => ({
@@ -47,7 +49,7 @@ vi.mock('../../../config.js', () => ({
 describe('listExtractRefreshTasksTool', () => {
   beforeEach(() => {
     vi.clearAllMocks();
-    mocks.mockAssertAdmin.mockResolvedValue(new Ok(true));
+    mocks.mockAssertAdmin.mockResolvedValue(undefined);
     mocks.mockQueryUserOnSite.mockResolvedValue({ siteRole: 'SiteAdministratorCreator' });
   });
